refactor: useCase package 생성
