+++ src/main/java/kr/hhplus/be/server/coupon/application/useCase/CreateCouponUseCase.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.coupon.application;
+package kr.hhplus.be.server.coupon.application.useCase;
 
 import kr.hhplus.be.server.coupon.application.dto.CouponRequest;
 import kr.hhplus.be.server.coupon.application.dto.CouponResponse;
