+++ src/test/java/kr/hhplus/be/server/coupon/application/CreateCouponUseCaseTest.java
@@ -2,6 +2,7 @@ package kr.hhplus.be.server.coupon.application;
 
 import kr.hhplus.be.server.coupon.application.dto.CouponRequest;
 import kr.hhplus.be.server.coupon.application.dto.CouponResponse;
+import kr.hhplus.be.server.coupon.application.useCase.CreateCouponUseCase;
 import kr.hhplus.be.server.coupon.domain.entity.Coupon;
 import kr.hhplus.be.server.coupon.domain.entity.CouponType;
 import kr.hhplus.be.server.coupon.infra.repositpry.port.CouponRepository;
